moved 'trust_transfer_limit' from safe to db
