+++ src/safe_db/db.rs
@@ -1,3 +1,4 @@
+use std::cmp::min;
 use std::collections::BTreeMap;
 
 use crate::types::{edge::EdgeDB, Address, Edge, Safe, U256};
@@ -36,8 +37,9 @@ impl DB {
                 if *user == *send_to {
                     continue;
                 }
+
                 if let Some(receiver_safe) = self.safes.get(send_to) {
-                    let limit = safe.trust_transfer_limit(receiver_safe, *percentage);
+                    let limit = self.trust_transfer_limit(safe, receiver_safe, *percentage);
                     if limit != U256::from(0) {
                         edges.push(Edge {
                             from: *user,
@@ -63,5 +65,47 @@ impl DB {
             }
         }
         self.edges = EdgeDB::new(edges)
+    }
+
+    /// This method calculates how much of their own tokens a user can send to a given receiver.
+    /// The transfer limit is based on the trust relationship between the user and the receiver,
+    /// denoted by the trust_percentage argument.
+    /// If the receiver is an organization, the method simply returns the balance of the sender, i.e.,
+    /// the user can send all their tokens to an organization.
+    /// If the receiver is a regular user, the method calculates the transfer limit based on the balance
+    /// of the receiver and the trust percentage. The method scales down the receiver's balance based
+    /// on the trust percentage and compares it with the original balance to calculate the maximum
+    /// amount the user can send.
+    /// The maximum amount a user can send is the smaller of the user's balance and the difference between
+    /// the scaled receiver's balance and the balance calculated based on the trust percentage.
+    /// @returns how much of their own tokens a user can send to receiver.
+    fn trust_transfer_limit(&self
+                            , sender: &Safe
+                            , receiver: &Safe
+                            , trust_percentage: u8) -> U256 {
+        if receiver.organization {
+            // TODO treat this as "return to owner"
+            // i.e. limited / only constrained by the balance edge.
+            sender.balance(&sender.token_address)
+        } else {
+            let receiver_balance = receiver.balance(&sender.token_address);
+
+            let amount = (receiver.balance(&receiver.token_address)
+                * U256::from(trust_percentage as u128))
+                / U256::from(100);
+
+            let scaled_receiver_balance =
+                receiver_balance * U256::from((100 - trust_percentage) as u128) / U256::from(100);
+
+            if amount < receiver_balance {
+                U256::from(0)
+            } else {
+                // TODO it should not be "min" - the second constraint is set by the balance edge.
+                min(
+                    amount - scaled_receiver_balance,
+                    sender.balance(&sender.token_address),
+                )
+            }
+        }
     }
 }
